third_party/blink/web_tests/platform/generic/fast/spatial-navigation/snav-symmetrically-positioned-expected.txt

This test ensures that Spatial Navigation works with symmetrically positoned rects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

E1
E2
E3
E4
E5
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e5"
PASS successfullyParsed is true

TEST COMPLETE
